fix bug permission

+++ app/Http/Controllers/ArticleController.php
@@ -71,8 +71,16 @@ class ArticleController extends Controller
     {
         try {
             $article = Article::findOrFail($id);
+            // dd(request()->user()->isAbleToAndOwns(['articles-update'], $article));
+            if (
+                request()->user()->hasRole(['superadmin', 'admin']) ||
+                request()->user()->isAbleToAndOwns(['articles-update'], $article)
+            ) {
+                return view('pages.article.edit', compact('article'));
+            } else {
+                return redirect()->route('article.index')->with($this->permissionDenied());
+            }
 
-            return view('pages.article.edit', compact('article'));
         } catch (ModelNotFoundException $e) {
             return redirect()->route('article.index')->with($this->alertNotFound());
         }
@@ -90,15 +98,22 @@ class ArticleController extends Controller
         try {
             $article = Article::findOrFail($id);
 
-            $this->validate($request, [
-                'title' => ['required', 'string', 'max:255'],
-                'body' => ['required', 'string'],
-                'published' => ['required'],
-            ]);
+            if (
+                request()->user()->hasRole(['superadmin', 'admin']) ||
+                request()->user()->isAbleToAndOwns(['articles-update'], $article)
+            ) {
+                $this->validate($request, [
+                    'title' => ['required', 'string', 'max:255'],
+                    'body' => ['required', 'string'],
+                ]);
 
-            $article->update($request->all());
+                $article->update($request->all());
 
-            return redirect()->route('article.index')->with($this->alertUpdated());
+                return redirect()->route('article.index')->with($this->alertUpdated());
+            } else {
+                return redirect()->route('article.index')->with($this->permissionDenied());
+            }
+
         } catch (ModelNotFoundException $e) {
             return redirect()->route('article.index')->with($this->alertNotFound());
         }
@@ -114,9 +129,19 @@ class ArticleController extends Controller
     {
         try {
             $article = Article::findOrFail($id);
-            $article->delete();
 
-            return redirect()->route('article.index')->with($this->alertDeleted());
+            if (
+                request()->user()->hasRole('superadmin') ||
+                request()->user()->isAbleToAndOwns('articles-delete', $article)
+            ) {
+
+                $article->delete();
+
+                return redirect()->route('article.index')->with($this->alertDeleted());
+            } else {
+                return redirect()->route('article.index')->with($this->permissionDenied());
+            }
+
         } catch (ModelNotFoundException $e) {
             return redirect()->route('article.index')->with($this->alertNotFound());
         }
